+++ ArcdpsLogManager/Logs/Filters/Composition/PlayerCountFilterType.cs
@@ -1,0 +1,9 @@
+namespace GW2Scratch.ArcdpsLogManager.Logs.Filters.Composition
+{
+	public enum PlayerCountFilterType
+	{
+		GreaterOrEqual,
+		Equal,
+		LessOrEqual
+	}
+}