+++ ArcdpsLogManager/Logs/Filters/ILogFilter.cs
@@ -5,7 +5,7 @@ namespace GW2Scratch.ArcdpsLogManager.Logs.Filters
 		/// <summary>
 		/// Determine whether a log should be kept or filtered out.
 		/// </summary>
-		/// <returns>A value indicating whether a log is kept (true) or filtered out (false).</returns>
+		/// <returns>A value indicating whether a log is kept (<see langword="true"/>) or filtered out (<see langword="false"/>).</returns>
 		bool FilterLog(LogData log);
 	}
 }